requirements.txt
fastapi==0.115.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.13.3
passlib==1.7.4
python-jose[cryptography]==3.3.0  
psycopg2==2.9.10
uvicorn==0.32.0
pydantic[email]==2.6.3  
python-multipart==0.0.12
bcrypt==4.2.0
